requirements.txt
albumentations==1.3.1
matplotlib==3.7.3
MulticoreTSNE==0.1
nibabel==5.2.1
numpy==1.24.3
opencv_python==4.8.1.78
opencv_python_headless==4.8.1.78
Pillow==10.0.1
scikit_learn==1.3.2
scipy==1.13.0
SimpleITK==2.3.1
scikit-image==0.21.0
tensorboardX==2.6.2.2
torch==1.12.1+cu113
torchvision==0.13.1+cu113
tqdm==4.66.1
tsnecuda==3.0.1
XlsxWriter==3.1.9
